Test UTF-8 behavior (already working in the app).
There was a bug preventing UTF-8 form data from being passed through to
the app correctly. It seemed to be in Plack::Test or maybe in Plack
itself. As it turns out, it was in URI::query_form(), which was being
called from POST() to construct the POST request. Encoding the form data
using URI::Escape::uri_escape() and assembling the request in the test
addresses the issue.

No changes were needed to the code under test.